2.2

+++ recipes/install.rb
@@ -2,25 +2,10 @@ require 'json'
 require 'base64'
 require 'digest'
 
-my_ip = my_private_ip()
 domain_name="domain1"
 domains_dir = node['hopsworks']['domains_dir']
 theDomain="#{domains_dir}/#{domain_name}"
-mysql_user=node['mysql']['user']
-mysql_password=node['mysql']['password']
 password_file = "#{theDomain}_admin_passwd"
-
-featurestore_user=node['featurestore']['user']
-featurestore_password=node['featurestore']['password']
-
-mysql_host = private_recipe_ip("ndb","mysqld")
-featurestore_jdbc_url = node['featurestore']['jdbc_url']
-# In case of an upgrade, attribute-driven-domain will not run but we still need to configure
-# connection pool for the online featurestore
-if node['featurestore']['jdbc_url'].eql? "localhost"
-  featurestore_jdbc_url="jdbc:mysql://#{mysql_host}:#{node['ndb']['mysql_port']}/"
-end
-
 
 bash "systemd_reload_for_glassfish_failures" do
   user "root"
@@ -43,18 +28,6 @@ group node['hopsworks']['group'] do
   not_if { node['install']['external_users'].casecmp("true") == 0 }
 end
 
-group node['jupyter']['group'] do
-  action :create
-  not_if "getent group #{node['jupyter']['group']}"
-  not_if { node['install']['external_users'].casecmp("true") == 0 }
-end
-
-group node['serving']['group'] do
-  action :create
-  not_if "getent group #{node['serving']['group']}"
-  not_if { node['install']['external_users'].casecmp("true") == 0 }
-end
-
 #
 # hdfs superuser group is 'hdfs'
 #
@@ -65,7 +38,7 @@ group node['hops']['hdfs']['user'] do
 end
 
 user node['hopsworks']['user'] do
-  home "/home/#{node['hopsworks']['user']}"
+  home node['glassfish']['user-home']
   gid node['hopsworks']['group']
   action :create
   shell "/bin/bash"
@@ -74,25 +47,12 @@ user node['hopsworks']['user'] do
   not_if { node['install']['external_users'].casecmp("true") == 0 }
 end
 
-group node['jupyter']['group'] do
-  action :modify
-  members ["#{node['hopsworks']['user']}"]
+group node["kagent"]["certs_group"] do
+  action :manage
   append true
+  excluded_members node['hopsworks']['user']
   not_if { node['install']['external_users'].casecmp("true") == 0 }
-end
-
-group node['serving']['group'] do
-  action :modify
-  members ["#{node['hopsworks']['user']}"]
-  append true
-  not_if { node['install']['external_users'].casecmp("true") == 0 }
-end
-
-group node['jupyter']['group'] do
-  action :modify
-  members ["#{node['hopsworks']['user']}"]
-  append true
-  not_if { node['install']['external_users'].casecmp("true") == 0 }
+  only_if { conda_helpers.is_upgrade }
 end
 
 group node['conda']['group'] do
@@ -110,43 +70,37 @@ group node['hops']['hdfs']['user'] do
   not_if { node['install']['external_users'].casecmp("true") == 0 }
 end
 
-user node['jupyter']['user'] do
-  home node['jupyter']['base_dir']
-  gid node['jupyter']['group']
+group node['kagent']['userscerts_group'] do
   action :create
-  shell "/bin/bash"
-  manage_home true
-  not_if "getent passwd #{node['jupyter']['user']}"
+  not_if "getent group #{node['kagent']['userscerts_group']}"
   not_if { node['install']['external_users'].casecmp("true") == 0 }
 end
 
-user node['serving']['user'] do
-  gid node['serving']['group']
-  action :create
-  shell "/bin/bash"
-  manage_home true
-  not_if "getent passwd #{node['serving']['user']}"
+group node['kagent']['userscerts_group'] do
+  action :modify
+  members node['hopsworks']['user']
+  append true
   not_if { node['install']['external_users'].casecmp("true") == 0 }
 end
 
-group node['kagent']['certs_group'] do
+group node['hops']['group'] do
+  gid node['hops']['group_id']
+  action :create
+  not_if "getent group #{node['hops']['group']}"
+  not_if { node['install']['external_users'].casecmp("true") == 0 }
+end
+
+group node['hops']['group'] do
   action :modify
   members ["#{node['hopsworks']['user']}"]
   append true
   not_if { node['install']['external_users'].casecmp("true") == 0 }
 end
 
-group node['hops']['group'] do
-  action :modify
-  members ["#{node['hopsworks']['user']}", "#{node['jupyter']['user']}", "#{node['serving']['user']}"]
-  append true
-  not_if { node['install']['external_users'].casecmp("true") == 0 }
-end
-
 #update permissions of base_dir to 770
 directory node['jupyter']['base_dir']  do
-  owner node['jupyter']['user']
-  group node['jupyter']['group']
+  owner node['hops']['yarnapp']['user']
+  group node['hops']['group']
   mode "770"
   action :create
 end
@@ -176,9 +130,35 @@ when "debian"
   if node['platform_version'].to_f <= 14.04
     node.override['hopsworks']['systemd'] = "false"
   end
-  package ["dtrx", "libkrb5-dev"]
+#  package ["dtrx", "libkrb5-dev"]
+  package ["p7zip-full", "libkrb5-dev"]
 
-  dtrx="dtrx"
+  remote_file "#{Chef::Config['file_cache_path']}/dtrx.tar.gz" do
+    user node['glassfish']['user']
+    group node['glassfish']['group']
+    source node['dtrx']['download_url']
+    mode 0755
+    action :create
+  end
+
+  bash "unpack_dtrx" do
+    user "root"
+    code <<-EOF
+      set -e
+      cd #{Chef::Config['file_cache_path']}
+      tar -xzf dtrx.tar.gz
+      cd dtrx-7.1
+      python setup.py install --prefix=/usr/local
+      # dtrx expects 7z to on its path. create a symbolic link from /bin/7z to /bin/7za
+      rm -f /bin/7z
+      ln -s /bin/7za /bin/7z
+    EOF
+    not_if "which dtrx"
+  end
+
+  dtrx="/usr/local/bin/dtrx"
+# dtrx="dtrx"
+
 when "rhel"
   package ["krb5-libs", "p7zip"]
 
@@ -314,27 +294,40 @@ node.override = {
             'maximumpoolsize' => 400,
             'taskqueuecapacity' => 20000,
             'description' => 'Hopsworks Executor Service'
+          },
+          'concurrent/condaExecutorService' => {
+              'threadpriority' => 9,
+              'corepoolsize' => 30,
+              'maximumpoolsize' => 400,
+              'taskqueuecapacity' => 20000,
+              'description' => 'Hopsworks Conda Executor Service'
           }
         },
         'managed_scheduled_executor_services' => {
           'concurrent/hopsScheduledExecutorService' => {
             'corepoolsize' => 10,
             'description' => 'Hopsworks Executor Service'
+          },
+          'concurrent/condaScheduledExecutorService' => {
+              'corepoolsize' => 10,
+              'description' => 'Hopsworks Conda Executor Service'
           }
         },
         'jdbc_connection_pools' => {
           'hopsworksPool' => {
             'config' => {
-              'datasourceclassname' => 'com.mysql.jdbc.jdbc2.optional.MysqlDataSource',
+              'datasourceclassname' => 'com.mysql.cj.jdbc.MysqlDataSource',
               'restype' => 'javax.sql.DataSource',
               'isconnectvalidatereq' => 'true',
               'validationmethod' => 'auto-commit',
               'ping' => 'true',
               'description' => 'Hopsworks Connection Pool',
               'properties' => {
-                'Url' => "jdbc:mysql://#{my_ip}:3306/",
-                'User' => mysql_user,
-                'Password' => mysql_password
+                'Url' => "jdbc:mysql://127.0.0.1:3306/",
+                'User' => node['hopsworks']['mysql']['user'],
+                'Password' => node['hopsworks']['mysql']['password'],
+                'useSSL' => 'false',
+                'allowPublicKeyRetrieval' => 'true'
               }
             },
             'resources' => {
@@ -345,16 +338,18 @@ node.override = {
           },
           'featureStorePool' => {
             'config' => {
-              'datasourceclassname' => 'com.mysql.jdbc.jdbc2.optional.MysqlDataSource',
+              'datasourceclassname' => 'com.mysql.cj.jdbc.MysqlDataSource',
               'restype' => 'javax.sql.DataSource',
               'isconnectvalidatereq' => 'true',
               'validationmethod' => 'auto-commit',
               'ping' => 'true',
               'description' => 'FeatureStore Connection Pool',
               'properties' => {
-                'Url' => featurestore_jdbc_url,
-                'User' => featurestore_user,
-                'Password' => featurestore_password
+                'Url' => node['featurestore']['hopsworks_url'],
+                'User' => node['featurestore']['user'],
+                'Password' => node['featurestore']['password'],
+                'useSSL' => 'false',
+                'allowPublicKeyRetrieval' => 'true'
               }
             },
             'resources' => {
@@ -365,16 +360,18 @@ node.override = {
           },
           'airflowPool' => {
             'config' => {
-              'datasourceclassname' => 'com.mysql.jdbc.jdbc2.optional.MysqlDataSource',
+              'datasourceclassname' => 'com.mysql.cj.jdbc.MysqlDataSource',
               'restype' => 'javax.sql.DataSource',
               'isconnectvalidatereq' => 'true',
               'validationmethod' => 'auto-commit',
               'ping' => 'true',
               'description' => 'Airflow Connection Pool',
               'properties' => {
-                'Url' => "jdbc:mysql://#{my_ip}:3306/",
+                'Url' => "jdbc:mysql://127.0.0.1:3306/",
                 'User' => node['airflow']['mysql_user'],
-                'Password' => node['airflow']['mysql_password']
+                'Password' => node['airflow']['mysql_password'],
+                'useSSL' => 'false',
+                'allowPublicKeyRetrieval' => 'true'
               }
             },
             'resources' => {
@@ -385,17 +382,18 @@ node.override = {
           },
           'ejbTimerPool' => {
             'config' => {
-              'datasourceclassname' => 'com.mysql.jdbc.jdbc2.optional.MysqlDataSource',
+              'datasourceclassname' => 'com.mysql.cj.jdbc.MysqlDataSource',
               'restype' => 'javax.sql.DataSource',
               'isconnectvalidatereq' => 'true',
-
               'validationmethod' => 'auto-commit',
               'ping' => 'true',
-              'description' => 'Hopsworks Connection Pool',
+              'description' => 'Hopsworks EJB Connection Pool',
               'properties' => {
-                'Url' => "jdbc:mysql://#{my_ip}:3306/glassfish_timers",
-                'User' => mysql_user,
-                'Password' => mysql_password
+                'Url' => "jdbc:mysql://127.0.0.1:3306/glassfish_timers",
+                'User' => node['hopsworks']['mysql']['user'],
+                'Password' => node['hopsworks']['mysql']['password'],
+                'useSSL' => 'false',
+                'allowPublicKeyRetrieval' => 'true'
               }
             },
             'resources' => {
@@ -435,6 +433,16 @@ else
   end
 end 
 
+mysql_connector = File.basename(node['hopsworks']['mysql_connector_url'])
+
+remote_file "#{theDomain}/lib/#{mysql_connector}"  do
+  user node['glassfish']['user']
+  group node['glassfish']['group']
+  source node['hopsworks']['mysql_connector_url']
+  mode 0755
+  action :create_if_missing
+end
+
 cauth = File.basename(node['hopsworks']['cauth_url'])
 
 remote_file "#{theDomain}/lib/#{cauth}"  do
@@ -442,22 +450,6 @@ remote_file "#{theDomain}/lib/#{cauth}"  do
   group node['glassfish']['group']
   source node['hopsworks']['cauth_url']
   mode 0755
-  action :create_if_missing
-end
-
-template "#{theDomain}/docroot/404.html" do
-  source "404.html.erb"
-  owner node['glassfish']['user']
-  group node['glassfish']['group']
-  mode "444"
-  action :create
-end
-
-cookbook_file"#{theDomain}/docroot/hops_icon.png" do
-  source 'hops_icon.png'
-  owner node['glassfish']['user']
-  group node['glassfish']['group']
-  mode '0755'
   action :create_if_missing
 end
 
@@ -508,14 +500,9 @@ ulimit_domain node['hopsworks']['user'] do
   end
 end
 
-
-  hopsworks_grants "reload_systemd" do
-    tables_path  ""
-    views_path ""
-    rows_path  ""
-    action :reload_systemd
+  kagent_config "glassfish-domain1" do 
+    action :systemd_reload
   end
-
 end
 
 ca_dir = node['certs']['dir']
@@ -523,7 +510,7 @@ ca_dir = node['certs']['dir']
 directory ca_dir do
   owner node['glassfish']['user']
   group node['kagent']['certs_group']
-  mode "750"
+  mode "755"
   action :create
 end
 
@@ -538,7 +525,7 @@ end
 
 directory "#{ca_dir}/transient" do
   owner node['glassfish']['user']
-  group node['kagent']['certs_group']
+  group node['kagent']['userscerts_group']
   mode "750"
   action :create
 end
@@ -602,6 +589,22 @@ kagent_sudoers "jupyter" do
   not_if       { node['install']['kubernetes'].casecmp("true") == 0 }
 end
 
+kagent_sudoers "convert-ipython-notebook" do 
+  user          node['glassfish']['user']
+  group         "root"
+  script_name   "convert-ipython-notebook.sh"
+  template      "convert-ipython-notebook.sh.erb"
+  run_as        "ALL" # run this as root - inside we change to different users 
+end
+
+kagent_sudoers "tensorboard" do 
+  user          node['glassfish']['user']
+  group         "root"
+  script_name   "tensorboard.sh"
+  template      "tensorboard.sh.erb"
+  run_as        "ALL" # run this as root - inside we change to different users 
+end
+
 kagent_sudoers "tfserving" do 
   user          node['glassfish']['user']
   group         "root"
@@ -646,14 +649,6 @@ kagent_sudoers "ca-keystore" do
   only_if       { node['hopsworks']['dela']['enabled'].casecmp("true") == 0 }
 end
 
-kagent_sudoers "start-llap" do 
-  user          node['glassfish']['user']
-  group         node['hive2']['group']
-  script_name   "start-llap.sh"
-  template      "start-llap.sh.erb"
-  run_as        node["hive2"]['user']
-end
-
 command=""
 case node['platform']
  when 'debian', 'ubuntu'
@@ -665,7 +660,7 @@ end
 template "#{theDomain}/bin/tfserving-launch.sh" do
   source "tfserving-launch.sh.erb"
   owner node['glassfish']['user']
-  group node['serving']['group']
+  group node['glassfish']['group']
   mode "550"
   variables({
      :command => command
@@ -686,9 +681,9 @@ template "#{theDomain}/bin/unzip-hdfs-files.sh" do
   action :create
 end
 
-["convert-ipython-notebook.sh", "anaconda-rsync.sh", "zip-hdfs-files.sh", "zip-background.sh",
-  "unzip-background.sh", "anaconda-command-ssh.sh", "conda-command-ssh.sh", "tensorboard.sh", "tensorboard-launch.sh", 
-  "tensorboard-cleanup.sh", "condasearch.sh", "pipsearch.sh", "list_environment.sh"].each do |script|
+["zip-hdfs-files.sh", "zip-background.sh", "unzip-background.sh",  "tensorboard-launch.sh",
+ "tensorboard-cleanup.sh", "condasearch.sh", "list_environment.sh", "jupyter-kill.sh",
+ "jupyter-launch.sh", "tfserving-kill.sh", "sklearn_serving-launch.sh", "sklearn_serving-kill.sh"].each do |script|
   template "#{theDomain}/bin/#{script}" do
     source "#{script}.erb"
     owner node['glassfish']['user']
@@ -698,21 +693,12 @@ end
   end
 end
 
-["jupyter-kill.sh", "jupyter-launch.sh"].each do |script|
-  template "#{theDomain}/bin/#{script}" do
-    source "#{script}.erb"
-    owner node['glassfish']['user']
-    group node['jupyter']['group']
-    mode "750"
-    action :create
-  end
-end
 
-["tfserving-kill.sh", "sklearn_flask_server.py", "sklearn_serving-launch.sh", "sklearn_serving-kill.sh"].each do |script|
+["sklearn_flask_server.py"].each do |script|
   template "#{theDomain}/bin/#{script}" do
     source "#{script}.erb"
     owner node['glassfish']['user']
-    group node['serving']['group']
+    group node['glassfish']['group']
     mode "750"
     action :create
   end
@@ -750,13 +736,14 @@ end
 # Hopsworks will use a sudoer script to launch jupyter as the 'jupyter' user.
 # The jupyter user will be able to read the files and write to the directories due to group permissions
 
-user node["jupyter"]["user"] do
-  home node["jupyter"]["base_dir"]
-  gid node["jupyter"]["group"]
-  action :create
-  shell "/bin/bash"
+user node['hops']['yarnapp']['user'] do
+  uid node['hops']['yarnapp']['uid']
+  gid node['hops']['group']
+  system true
   manage_home true
-  not_if "getent passwd #{node["jupyter"]["user"]}"
+  shell "/bin/bash"
+  action :create
+  not_if "getent passwd #{node['hops']['yarnapp']['user']}"
   not_if { node['install']['external_users'].casecmp("true") == 0 }
 end
 
@@ -815,19 +802,18 @@ bash "unpack_flyway" do
     mv #{flyway} #{theDomain}
     cd #{theDomain}
     chown -R #{node['glassfish']['user']} flyway*
-    rm -rf flyway
+    if [ -L flyway ]; then
+      cp -r flyway/sql #{flyway}/ 
+      rm -rf flyway
+    fi
     ln -s #{flyway} flyway
   EOF
-  not_if { ::File.exists?("#{theDomain}/flyway/flyway") }
 end
 
 template "#{theDomain}/flyway/conf/flyway.conf" do
   source "flyway.conf.erb"
   owner node['glassfish']['user']
   mode 0750
-  variables({
-              :mysql_host => my_ip
-            })
   action :create
 end
 
@@ -856,3 +842,20 @@ directory "#{theDomain}/flyway/dml/undo" do
   action :create
 end
 
+directory "#{theDomain}/flyway/all" do
+  owner node['glassfish']['user']
+  mode "770"
+  action :create
+end
+
+directory "#{theDomain}/flyway/all/sql" do
+  owner node['glassfish']['user']
+  mode "770"
+  action :create
+end
+
+directory "#{theDomain}/flyway/all/undo" do
+  owner node['glassfish']['user']
+  mode "770"
+  action :create
+end
